+++ contrib/brl/bbas/brad/brad_multispectral_functions.cxx
@@ -52,8 +52,8 @@ void brad_polygon2mask(const vil_image_view<float>& image,
   vil_image_view<bool>& mask) {
   mask.set_size(image.ni(), image.nj());
   mask.fill(false);
-  for (int y = 0; y < image.nj(); y++) {
-    for (int x = 0; x < image.ni(); x++) {
+  for (size_t y = 0; y < image.nj(); y++) {
+    for (size_t x = 0; x < image.ni(); x++) {
       if (poly.contains(x, y)) mask(x, y) = true;
     }
   }
@@ -78,8 +78,8 @@ bool brad_compute_normalized_index_image(
   if (mul_img.nplanes() != 8 && mul_img.nplanes() != 16) return false;
   idx.set_size(mul_img.ni(), mul_img.nj());
 
-  for (int y = 0; y < mul_img.nj(); y++) {
-    for (int x = 0; x < mul_img.ni(); x++) {
+  for (size_t y = 0; y < mul_img.nj(); y++) {
+    for (size_t x = 0; x < mul_img.ni(); x++) {
       float denom = mul_img(x, y, b1) + mul_img(x, y, b2);
       if (fabs(denom) > tol)
         idx(x, y) = (mul_img(x, y, b1) - mul_img(x, y, b2)) / denom;
